[#ENYO-850] fix 'params' manipulation in the wrong order
createFile & deleteFile do weird manipulations on the parameters... transforming a signature which is otherwise homogeneous among other Hermes verbs.  One of these manipulation over-writes the next() value & hence loses the code that is supposed to write the HTTP response.  Restoring the next() value fixes this specific issue.

This is a tactical bug-fix, in the spirit of the original code.  Not the way I would have written this code.