Update for ENYO-1697 fuller fix
This allows the Designer to add in the fuller values from published and events properties from the original analyzed kind.  It doesn't add in any thing if the kind has not been analyzed yet or does not exist at all.  This does not change behavior for bringing changes made in designer to these values back into the Code Editor.