ENYO-1269: fix file save on Firefox
It is actually a work-around against a Firefox bug (tested Ok on FF17)

FormData#append() lacks the third parameter
'filename', so emulate it using a list of
'filename'fields of the same size os the
number of files.  This only works if the
other end of the tip is implemented on
server-side.
http://stackoverflow.com/questions/6664967/how-to-give-a-blob-uploaded-as-formdata-a-file-name

Enyo-DCO-1.1-Signed-off-by: Francois-Xavier KOWALSKI <francois-xavier.kowalski@hp.com>