Enyo-1779: Splash screen url is added to Android, IOS & Black Berry configuration
Phonegap build let the developper define multiple Splash screens adapted to the resolution of the device (Android and IOS only). This widget will be improved in the task Enyo-2433.

This commit include also some minor changes in the CSS style of the "Signing key" row attached to the panel "Project - Phonegap Build"