ENYO-2405: The download of the the built package from Phonegap is now implemented
This commit contain also a code refactoring for "BuildStatusUI.js" and a new feature which let the user access to the web page of "Phonegap build" by clicking on a status label. Due to the problem of timeout which occure when Ares tries to download a packaged application from Phonegap using the API-Phonegap-build, an alternative code is used, it allows Node.js to grap a file from any directory on the system and send it in a multipart/data form to the target directory. Note that the code using the API-Phonegap-Build is commented.