ENYO-2405: Fixed bug due to the merge.
The build and the buildStatus features work, the pull request remarks will be treated in the next commit.
Note: in this commit the loading pop-up displayed when the build status is requested continue to run after dismissing the Status pop-up, this bug will be also fixed in the next commit.