ENYO-2403: Refactoring of the data structure in "Phonegap.EditUiData" used to create the UI in Phonegap build panel
The creation of drawers in this UI is also refactored.
The button "Details" used to fold/unfold a drawer is removed, the trigger to do this action is the label of the drawers.
The general Icon input is positionned in the drawer general.
A problem in implementation cause the "keySelector" to not be instanciated in the drawers "Android", "IOS" and "BlackBerry", this will be correted in the next commits.