ENYO-2404: the "project.js" is now updated by the selected value of all the drawers row except thoses specific to the definition of the icon or the splash screen.
So, when the Project->Edit Popup is dissmissed by clicking on "Ok" the "project.js" is updated and when the build is launched, the "config.xml" is generated according to the selected parameters form the UI.
Also default value are displayed for each row of the UI except fo the icon rows and the splashScreen rows.
The actual version isn't stable yet, sometimes there is an error where the name of the project is replaced by "BUG IF YOU SEE THIS", this error will be fixed in one of the next commits.