ENYO-2591: the "None" element is no longer present in the siging key picker (Phonegap build panel).
Since it's not possible to disjoin a signing key from a Phonegap user, when a signing key is created it's no longer possible to build an application without selecting a signing key. To make the Phonegap build UI coherent with this constraint the "None" element is discared from the Signing keys picker.
Also in this commit, the signing keys row is rearranged vertically to be properly displayed.
In the case where there is no signing key associated with a platform in the user account, the message "Ne signing keys for this platform" is displayed instead of having an empty signing keys picker.