ENYO-3053: Show the Signing key title while loading Phonegap userData.
If a signing key is already defined for the project & Ares is waiting the userData object request then a message containing the name of the defined Signing Key is displayed in the row Signing Key.
If a signing key is not defined for the project & Ares is waiting the userData object request then a message containing only the loading message  is displayed in the row Signing Key.