ENYO-3133: Test if the interval min-max API level is correctelly set.
If the min value is greater thant the max value or vis versa, an error message is displayed under the two SDK version Pickers.
There is a bug accuring when there is an error on the SDK min/max version selection and the Splash screen duration or Load URL timeout input
then the correction of one of them enable the ok button of the Pop-up.
This bug will be corrected in one of the next commits.