ENYO-3061, ENYO-2514: stop blocking Ares while PhoneGap is building
Make the ZIP-file & uploading it to PhoneGap build still blocks the IDE
but the packages are now downloaded asynchronously.

When each is package is locally downloaded the `PhoneGap.Build` emits
an `onFsEvent` so that Ares refreshes the FileTree.

This commit also collapses 2 function names in one, for code readability.

Enyo-DCO-1.1-Signed-off-by: Francois-Xavier KOWALSKI <francois-xavier.kowalski@hp.com>