ENYO-3506: Use a static error to signal bukld cancelled
The build steps are executed in sequence using an Async.waterfall call. When the build is cancelled, the waterfall sequence needs to be stopped. The proper way to do that is to call the next() function with an Error() as parameter. So far, when a build generates an Error(), it gets propagated at project level so that the UI can display an error message. When the error is a build cancel, the Error should not be propagated as no message needs to be displayed.
The "build cancel" error is defined as a statics to make that sure everywhere we cancel (can happen during any step) we use the same specific Error, which can be detected easily on lines 496-498.