ENYO-3634: Ares proxy services are local, CORS should be filtered out
When services are being called through Ares, the client calls Ares first which act as a proxy to the service. If the request is a CORS request, Ares is to decide if it loet goes or not.
The service should not be exposed to whether the browser<->Ares relationship was same origin or not.
When passing on the request to the service, we remove any possible 'Origin' header of simple CORS requests. This will make any CORS request non-CORS (OPTIONS are never processed here).
When parsing the response, we remove any CORS headers that may have been set so that Ares does not get its own CORS headers to be replaced by what has been set by the service.