ENYO-3771: all phonegap API calls should contain the authentication token
The PGB token is stored as a cookie. When PGB service is used under CORS conditions (e.g. Ares is run on localhost:9009), cookie are by default *not sent* to server. The PGB service does not therefore receive the authentication token and operation fails.
We need to enable the xhr option "withCredentials". This allow cookies to be sent, which repair the faulty behavior.